chore: init v4819 script
